Moves createConstructor to Enum
